requirements.txt
streamlit
tensorflow==2.15.0
tf-keras
pillow
numpy